+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.4.0 (08 Dec, 2016)
+
+* Change return value of events to return an object
+* Checkbox/radio inputs now return an additional property `checked`
+
 ## 0.3.0 (03 Dec, 2016)
 
 * Add label widget